now using GenericComponent for VerdictPanel, color coding of pro/opp for the buttons, changed the padding on the debates detail page to have the components sit even. TESTS: POST works, but PATCH doesn't (for some reason), so I started to update the other tests, which also now don't work (for some reason???)
