+++ src/components/tournament/MarshalPanel.tsx
@@ -13,13 +13,13 @@ export function MarshalPanel({ motion }: MarshalPanelProps) {
     <GenericComponent
       title="Marshal panel"
       showActions
-      className="mt-8 w-full max-w-[760px] sm:mt-10"
+      className="w-full max-w-[760px]"
     >
       <a
         href={conductDebateHref}
         target="_blank"
         rel="noreferrer"
-        className="block w-full rounded-sm border border-stone-600 bg-white/10 px-6 py-5 text-center text-xl font-semibold text-white transition hover:bg-white/15"
+        className="block w-full rounded-sm border border-stone-600 bg-white/10 px-6 py-5 text-center text-xl font-semibold text-white/75 transition hover:bg-white/15"
       >
         Conduct debate!
       </a>
